Swiftora/backend/requirements.txt
fastapi>=0.110.0,<0.111.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pandas>=2.0
numpy>=1.22
pillow>=9.0
python-dotenv>=1.0
pydantic>=2.0
pytest>=7.0
python-dateutil>=2.8
scikit-learn>=1.2
rapidfuzz>=3.0